==> jupyter_notebooks/05-correlation-study.ipynb <==
# **Correlation Analysis Notebook**

## Objectives
- Perform a correlation study on the variables of the dataset, to see what categorical variable most strongly influence the numerical variables
- Perform a Predictive Power Score study on the variables of the dataset, to see what categoricals have the greatest predictive power for the numerical variable
- Partially fulfil Business Requirement 1

## Inputs
- outputs/datasets/collection/student-exam-results.csv

## Outputs

- A understanding of what categorical variables influence the numerical variables, and to what degree
- A understanding of what categorical variables have the greatest predictive power for the numerical variables

## Import Packages

import os
import pandas as pd
from feature_engine.encoding import OneHotEncoder
import ppscore as pps

# Change working directory

Since this notebook exists in the jupyter_notebooks directory, we need to change the current working directory from the jupyter_notebooks directory to the workspace, so that any directories created in further codes cells are added in the correct place. 

We access the current directory with the OS packages' `getcwd()` method

current_directory = os.getcwd()
current_directory


We now want to set the working directory as the parent of the current working directory, jupyter_notebooks

- The `os.path.dirname()` method gets the parent directory
- The `os.chir()` method defines the new current directory
- We do this to access all of the project's files and directories, rather than those in the jupyter_notebooks directory

os.chdir(os.path.dirname(current_directory))
print("You set a new current directory")

To make certain of things, we now use a code cell to confirm that we have set the current working directory properly

current_directory = os.getcwd()
current_directory

## Correlation Study

Now that we have become more familiar with the dataset, we can perform a correlation study. As is noted in the Churnometer walkthrough project, we could use the Predictive Power Score library, however the feature variables are categorical. This means that we use One Hot Encoding to encode the categorical variables so that we can determine correlation scores.

However, our task is larger in scope, since we have 5 target variables. We need to see how each feature variable correlates with each target variable. First, as we have a new notebook, we need to re-import the dataset.

df = pd.read_csv('outputs/datasets/collection/student-exam-results.csv')
df.head()

Now we can apply the One Hot Encoder

one_hot_encoder = OneHotEncoder(variables = df.columns[df.dtypes=='object'].to_list(), drop_last=False)
df_one_hot_encoder = one_hot_encoder.fit_transform(df)
print(df_one_hot_encoder.shape)
df_one_hot_encoder.head()

We note that the encoded dataset now has 21 columns, and that the 5 target variables have been preserved.

### Maths score correlation

#### Pearson

math_correlation_pearson = df_one_hot_encoder.corr(method='pearson')['math_score'].sort_values(key=abs, ascending=False)[1:].head(10)
math_correlation_pearson

We can see that the other 4 target variables correlate highly with math_score. By default, we exclude the first entry, which is math_score's correlation with itself. Since, in any prediction situation, we will not know the other test scores, we cannot use them in the correlation study, so we can adjust the correlation code cell to exclude the first 5 rows, so as to better reveal how the feature variables correlate with the math_score variable.

math_correlation_pearson = df_one_hot_encoder.corr(method='pearson')['math_score'].sort_values(key=abs, ascending=False)[5:].head(10)
math_correlation_pearson

This reveals that only the lunch_program feature variable correlates with the math_score variable, and then only weakly. Therefore, prediction of the math_score target variable could be problematic.

#### Spearman

We anticipate a similar outcome to the Pearson correlation, so we will discard the first 5 rows of the Spearman correlation test output.

math_correlation_spearman = df_one_hot_encoder.corr(method='spearman')['math_score'].sort_values(key=abs, ascending=False)[5:].head(10)
math_correlation_spearman

We get similar correlation results to the Pearson correlation - the lunch_program variable correlates weakly to the math_score variable.

### Reading score correlation

#### Pearson

reading_correlation_pearson = df_one_hot_encoder.corr(method='pearson')['reading_score'].sort_values(key=abs, ascending=False)[5:].head(10)
reading_correlation_pearson

We see that both the lunch_program and test_prepartion_course feature variables correlate weakly with reading_score

#### Spearman

reading_correlation_spearman = df_one_hot_encoder.corr(method='spearman')['reading_score'].sort_values(key=abs, ascending=False)[5:].head(10)
reading_correlation_spearman

As with Pearson correlation, we see that both the lunch_program and test_prepartion_course feature variables correlate weakly with reading_score

### Writing score correlation

#### Pearson

writing_correlation_pearson = df_one_hot_encoder.corr(method='pearson')['writing_score'].sort_values(key=abs, ascending=False)[5:].head(10)
writing_correlation_pearson

We see that lunch_program and test_preparation_course correlate weakly with writing_score

#### Spearman

reading_correlation_spearman = df_one_hot_encoder.corr(method='spearman')['writing_score'].sort_values(key=abs, ascending=False)[5:].head(10)
reading_correlation_spearman

We see that lunch_program and test_preparation_course correlate weakly with writing_score

## Discussion

We see that both the Spearman and Pearson correlation methods produce similar correlation scores. The Pearson method slightly outperforms the Spearman method. Both methods produce the same rankings of variables. The lunch_program variable correlates most highly with the test score variables, followed by the test_preparation_course variable.

## Predictive Power Score

As noted above, we chose to use One Hot Encoding and the correlation method to determine correlation scores. However, at best, we only saw weak correlations. So, we will also analyse the dataset with the Predictive Power Score module, to see if any further insights can be gathered.

Since we have seen in the correlation studies that the score variables tend to correlate well with each other, it is likely that the Predictive Power Score module will also assign strong predictive power to the score variables for the other score variables, which may confuse matters. Hence, we will run 3 predictive power score code cells - one for each of the score variables. To do this, we will generate dataframes that only contain one score variable.

First, lets recap the dataframe:

df.head()

### Maths

To assess which variables have predictive power for the math_score variable, we need to drop the reading_score, writing_score and average_score columns:

df_maths = df.drop(['reading_score', 'writing_score', 'average_score'], axis=1)
df_maths.head()

Now let's generate the Predictive Power Score dataframe

df_maths_pps = pps.matrix(df=df_maths)
df_maths_pps

This dataframe is unwieldy, and we are only interested in a few items, so let's filter and query it. We are not interested in the rows that do not deal with the math_score variable, nor are we interested in the columns after `ppscore`. 

The only item we will retain from the full dataframe is the methpd's choice of the DecisionTreeClassifier and DecisionTreRegressor algorithms. This may prove relevant when we come to fit a prediction model.

df_maths_pps_filtered = df_maths_pps.filter(['x', 'y', 'ppscore']).query('x == "math_score"')
df_maths_pps_filtered

As noted in [this forum discussion](https://github.com/8080labs/ppscore/issues/39) on PPS' Github page, a score of >0.2 indicates strong predictive power, and a score of 0 to 2 indicates some weak yet relevant predictive power.

As we can see, the PPS module has determined that lunch_program has strong predictive power for a student's math score. This is in line with the outcome of the correlation study. The other variables have very low predictive power scores, and we can safely discard them. The test_preparation_course variable also has practically no predictive power, which is in line with our previous correlation study.

### Reading

As before, we'll filter the initial dataframe so that the only numerical variable is reading_score.

df_reading = df.drop(['math_score', 'writing_score', 'average_score'], axis=1)
df_reading.head()

Now the predictive power score dataframe:

df_reading_pps = pps.matrix(df=df_reading)
df_reading_pps

As before, this is cluttered, so we'll filter and query. We will also mark the use of use the DecisionTreeClassifier algorithm 

df_reading_pps_filtered = df_reading_pps.filter(['x', 'y', 'ppscore']).query('x == "reading_score"')
df_reading_pps_filtered

Interestingly, lunch_program has slightly less predictive power for reading_score. The test_preparation_course variable has also assumed slightly higher predictive power than for math_score. Given that interpretation of predictive power score is subjective and contextual, we can probably take lunch_program as having strong predictive power, given that it is the highest score by far. It is interesting that test_preparation_course has such a low predictive power score when it correlates only a little less strongly than lunch_program (0.24 vs 0.28).

### Writing

As above, we will first filter the initial dataframe:

df_writing = df.drop(['math_score', 'reading_score', 'average_score'], axis=1)
df_writing.head()

Now the predictive power score dataframe:

df_writing_pps = pps.matrix(df=df_writing)
df_writing_pps

As above, we need to filter and query. Again, we note the use of the DecisionTreeClassifier algorithm

df_writing_pps_filtered = df_writing_pps.filter(['x', 'y', 'ppscore']).query('x == "writing_score"')
df_writing_pps_filtered

We note that lunch_program continues to enjoy strong predictive power, and that test_preparation_course also has a high predictive power score. As with reading_score, we will use the contextual nature of PPS interpretation to our advantage and classify it as strong predictive power. This may be justified by the very close correlation scores.

### Discussion

We will gather our insights from the predictive power score analyses here:

- For math_score, lunch_program has the strongest predictive power by far

- For reading_score, lunch_program has the strongest predictive power

- For writing_score, lunch_program has the strongest predictive power, closely followed by test_preparation_course

Therefore, going forward, we will focus on the lunch_program variable and, to a lesser extent, the test_preparation_course variable.
