==> jupyter_notebooks/07-DataSplitting.ipynb <==
# **Data Splitting**

## Objectives
- Split the dataset into train and test sets for each of the maths, writing and reading exam scores, so that models can be trained to predict each exam score

## Inputs
- outputs/datasets/collection/student-exam-results.csv

## Outputs
- train and test sets for each of the math_score, reading_score and writing_score variables

## Additional comments

# Change working directory

Since this notebook exists in the jupyter_notebooks directory, we need to change the current working directory from the jupyter_notebooks directory to the workspace, so that any directories created in further codes cells are added in the correct place. 

We access the current directory with the OS packages' `getcwd()` method

import os
current_directory = os.getcwd()
current_directory


We now want to set the working directory as the parent of the current working directory, jupyter_notebooks

- The `os.path.dirname()` method gets the parent directory
- The `os.chir()` method defines the new current directory
- We do this to access all of the project's files and directories, rather than those in the jupyter_notebooks directory

os.chdir(os.path.dirname(current_directory))
print("You set a new current directory")

To make certain of things, we now use a code cell to confirm that we have set the current working directory properly

current_directory = os.getcwd()
current_directory

## Load data

import pandas as pd
df = pd.read_csv('outputs/datasets/collection/student-exam-results.csv')
df.head()

Now we create three different sets

df_math_score = df.drop(['reading_score', 'writing_score', 'average_score', 'literacy_score'], axis=1)
df_math_score.head()

df_reading_score = df.drop(['math_score', 'writing_score', 'average_score', 'literacy_score'], axis=1)
df_reading_score.head()

df_writing_score = df.drop(['math_score', 'reading_score', 'average_score', 'literacy_score'], axis=1)
df_writing_score.head()

### Math Score

Now we can create split these datasets to form train and test sets. First, we import the train_test_split package.

from sklearn.model_selection import train_test_split

math_train_vars, math_test_vars, math_train_score, math_test_score = train_test_split(
    df_math_score.drop(['math_score'], axis=1),
    df_math_score['math_score'],
    test_size = 0.2,
    random_state = 7
)

Now we check to see if this has worked:

print(f"math_train_vars: {math_train_vars.shape}")
print(f"math_test_vars: {math_test_vars.shape}")
print(f"math_train_score: {math_train_score.shape}")
print(f"math_test_score: {math_test_score.shape}")


Excellent - we see that the math_train_vars and math_train_score datasets both have 800 entries, equivalent to 80% of the original 1000 record dataset. We also see that the math_test_vars and math_test_score datasets both have 200 entries, equivalent to 20% of the original 1000 record dataset. For good measure, we will inspect the dataframes and series below:

math_train_vars

math_test_vars

math_train_score

math_test_score

Upon inspection, we see that the data has also been randomised.

### Reading Score

We will now do the same with the reading_score variable

reading_train_vars, reading_test_vars, reading_train_score, reading_test_score = train_test_split(
    df_reading_score.drop(['reading_score'], axis=1),
    df_reading_score['reading_score'],
    test_size = 0.2,
    random_state = 7
)

Now we check to see if these have been created properly

print(f"reading_train_vars: {reading_train_vars.shape}")
print(f"reading_test_vars: {reading_test_vars.shape}")
print(f"reading_train_score: {reading_train_score.shape}")
print(f"reading_test_score: {reading_test_score.shape}")

Excellent - 800 records for the train set and 200 records for the test set. Since we have used the same random state, we can be assured that the data has been randomised

### Writing Score

Finally, we split the writing_score dataset

writing_train_vars, writing_test_vars, writing_train_score, writing_test_score = train_test_split(
    df_writing_score.drop(['writing_score'], axis=1),
    df_writing_score['writing_score'],
    test_size = 0.2,
    random_state = 7
)

print(f"writing_train_vars: {writing_train_vars.shape}")
print(f"writing_test_vars: {writing_test_vars.shape}")
print(f"writing_train_score: {writing_train_score.shape}")
print(f"writing_test_score: {writing_test_score.shape}")

As before, we have 800 records for the train set and 200 for the test set

### Save files

We can now save these various datasets as CSV files in the outputs folder. First, we'll create the necessary folders:

import os
try:
    ! rm -r outputs/datasets/filtered
    ! rm -r outputs/datasets/split
    os.makedirs(name='outputs/datasets/filtered')
    print('directory outputs/datasets/filtered created')
    os.makedirs(name='outputs/datasets/split')
    print('directory outputs/datasets/split created')
except Exception as e:
    print(e)

Now we save the filtered datasets, before any splitting happens

! rm outputs/datasets/filtered/math-score-dataset.csv
df_math_score.to_csv("outputs/datasets/filtered/math-score-dataset.csv", index=False)
print('math-score-dataset.csv created in outputs/datasets/filtered')

! rm outputs/datasets/filtered/reading-score-dataset.csv
df_reading_score.to_csv("outputs/datasets/filtered/reading-score-dataset.csv", index=False)
print('reading-score-dataset.csv created in outputs/datasets/filtered')

! rm outputs/datasets/filtered/writing-score-dataset.csv
df_writing_score.to_csv("outputs/datasets/filtered/writing-score-dataset.csv", index=False)
print('writing-score-dataset.csv created in outputs/datasets/filtered')

Now we can save the split datasets

try:
    
    ! rm -r outputs/datasets/split/math
    ! rm -r outputs/datasets/split/reading
    ! rm -r outputs/datasets/split/writing
    os.makedirs(name='outputs/datasets/split/math')
    print('directory outputs/datasets/split/math created')
    os.makedirs(name='outputs/datasets/split/reading')
    print('directory outputs/datasets/split/reading created')
    os.makedirs(name='outputs/datasets/split/writing')
    print('directory outputs/datasets/split/writing created')
except Exception as e:
    print(e)

! rm outputs/datasets/split/math/math-train-vars.csv
math_train_vars.to_csv("outputs/datasets/split/math/math-train-vars.csv", index=False)
print('math-train-vars.csv created in outputs/datasets/split/math')

! rm outputs/datasets/split/math/math-test-vars.csv
math_test_vars.to_csv("outputs/datasets/split/math/math-test-vars.csv", index=False)
print('math-test-vars.csv created in outputs/datasets/split/math')

! rm outputs/datasets/split/math/math-train-score.csv
math_train_score.to_csv("outputs/datasets/split/math/math-train-score.csv", index=False)
print('math-train-score.csv created in outputs/datasets/split/math')

! rm outputs/datasets/split/math/math-test-score.csv
math_test_score.to_csv("outputs/datasets/split/math/math-test-score.csv", index=False)
print('math-test-score.csv created in outputs/datasets/split/math')

! rm outputs/datasets/split/reading/reading-train-vars.csv
reading_train_vars.to_csv("outputs/datasets/split/reading/reading-train-vars.csv", index=False)
print('reading-train-vars.csv created in outputs/datasets/split/reading')

! rm outputs/datasets/split/reading/reading-test-vars.csv
reading_test_vars.to_csv("outputs/datasets/split/reading/reading-test-vars.csv", index=False)
print('math-test-vars.csv created in outputs/datasets/split/math')

! rm outputs/datasets/split/reading/reading-train-score.csv
reading_train_score.to_csv("outputs/datasets/split/reading/reading-train-score.csv", index=False)
print('reading-train-score.csv created in outputs/datasets/split/reading')

! rm outputs/datasets/split/reading/reading-test-score.csv
reading_test_score.to_csv("outputs/datasets/split/reading/reading-test-score.csv", index=False)
print('reading-test-score.csv created in outputs/datasets/split/reading')

! rm outputs/datasets/split/writing/writing-train-vars.csv
writing_train_vars.to_csv("outputs/datasets/split/writing/writing-train-vars.csv", index=False)
print('writing-train-vars.csv created in outputs/datasets/split/writing')

! rm outputs/datasets/split/writing/writing-test-vars.csv
writing_test_vars.to_csv("outputs/datasets/split/writing/writing-test-vars.csv", index=False)
print('writing-test-vars.csv created in outputs/datasets/split/writing')

! rm outputs/datasets/split/writing/writing-train-score.csv
writing_train_score.to_csv("outputs/datasets/split/writing/writing-train-score.csv", index=False)
print('writing-train-score.csv created in outputs/datasets/split/writing')

! rm outputs/datasets/split/writing/writing-test-score.csv
writing_test_score.to_csv("outputs/datasets/split/writing/writing-test-score.csv", index=False)
print('writing-test-score.csv created in outputs/datasets/split/writing')
